Fix link to a `packages.el` file
It seems that the file `spacemacs/packages.el` was eventually moved to
the `layers/+distribution/spacemacs-base/` directory. That is, *pieces*
of it, since the content has in any case been modified a lot, maybe some
things have been factored and moved into other files, etc. In any case,
it seems that linking to this file achieves the same effect as what was
originally intended.

Also change how the link to the file is referred to, since the absolute
path `spacemacs/packages.el` doesn’t make sense any more.